prometheus_v3/requirements.txt
# ============================================================================
# PROMETHEUS V3 - REQUIREMENTS
# Python dependencies for production
# ============================================================================

# ============================================================================
# Core Dependencies
# ============================================================================

# Async and concurrency
asyncio-mqtt==0.16.2
aiohttp==3.9.1
aiofiles==23.2.1
uvloop==0.19.0

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# ============================================================================
# AI Providers
# ============================================================================

# OpenAI
openai==1.3.7

# Anthropic Claude
anthropic==0.7.7

# Google
google-generativeai==0.3.1
google-auth==2.25.2
google-cloud-aiplatform==1.38.1

# ============================================================================
# Database and Storage
# ============================================================================

# PostgreSQL
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.0

# Redis
redis==5.0.1
aioredis==2.0.1

# Vector Database
chromadb==0.4.18
qdrant-client==1.7.0

# ============================================================================
# Web Automation
# ============================================================================

# Playwright
playwright==1.40.0

# Selenium fallback
selenium==4.16.0
webdriver-manager==4.0.1

# BeautifulSoup for parsing
beautifulsoup4==4.12.2
lxml==4.9.3

# ============================================================================
# NLP and Text Processing
# ============================================================================

# spaCy
spacy==3.7.2
# Note: Download model separately with: python -m spacy download pt_core_news_sm

# NLTK
nltk==3.8.1

# Text processing
ftfy==6.1.3
unidecode==1.3.7

# ============================================================================
# Computer Vision
# ============================================================================

# OpenCV
opencv-python-headless==4.8.1.78

# Pillow
Pillow==10.1.0

# OCR
pytesseract==0.3.10
easyocr==1.7.1

# ============================================================================
# Audio Processing
# ============================================================================

# Speech recognition
SpeechRecognition==3.10.1
pyaudio==0.2.14

# Text to speech
pyttsx3==2.90
gtts==2.5.0

# Audio processing
pydub==0.25.1

# ============================================================================
# Data Processing
# ============================================================================

# Pandas and NumPy
pandas==2.1.4
numpy==1.24.3

# Data validation
pydantic==2.5.2
pydantic-settings==2.1.0

# ============================================================================
# API and Integrations
# ============================================================================

# HTTP clients
httpx==0.25.2
requests==2.31.0

# WhatsApp
yowsup==3.3.0

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# ============================================================================
# Scheduling and Tasks
# ============================================================================

# APScheduler
apscheduler==3.10.4

# Celery
celery==5.3.4
flower==2.0.1

# ============================================================================
# Monitoring and Logging
# ============================================================================

# Structured logging
structlog==23.2.0
python-json-logger==2.0.7

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0

# ============================================================================
# Security
# ============================================================================

# Cryptography
cryptography==41.0.7
pycryptodome==3.19.0

# JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ============================================================================
# Development Tools
# ============================================================================

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.3

# Debugging
ipython==8.18.1
ipdb==0.13.13

# ============================================================================
# File Processing
# ============================================================================

# YAML
pyyaml==6.0.1

# PDF
PyPDF2==3.0.1
pdfplumber==0.10.3

# Excel
openpyxl==3.1.2
xlsxwriter==3.1.9

# ============================================================================
# System and Utils
# ============================================================================

# System monitoring
psutil==5.9.6

# Environment management
python-dotenv==1.0.0

# File watching
watchdog==3.0.0

# Date and time
python-dateutil==2.8.2
pytz==2023.3

# Rich terminal output
rich==13.7.0
click==8.1.7

# ============================================================================
# Machine Learning (Optional)
# ============================================================================

# scikit-learn==1.3.2
# transformers==4.36.2
# torch==2.1.2
# tensorflow==2.15.0

# ============================================================================
# Additional Tools
# ============================================================================

# QR Code
qrcode[pil]==7.4.2

# Markdown
markdown==3.5.1

# JSON Schema
jsonschema==4.20.0

# Rate limiting
slowapi==0.1.9

# Retry logic
tenacity==8.2.3

# Progress bars
tqdm==4.66.1

# ============================================================================
# Version pinning for stability
# ============================================================================

# Pin setuptools and wheel
setuptools==69.0.2
wheel==0.42.0
